Escapes the milestone name before returning it from the macro
This prevents using milestone names for an XSS attack.

This is supposedly (although we aren't convinced) fixed in the 0.9
branch upstream, but it is definately not fixed in the 0.11 branch, the
one used in #define.  We may want to share this upstream, but the plugin
appears to be abandoned.

Fixes #4981

git-svn-id: file:///mnt/ramdisk/d4-svn/trunk@24730 484df14f-75d2-4b3f-8ba0-490e05069a01